Refactor server side into two classes
addresses issue #14
