+++ server/src/main/java/umm3601/Server.java
@@ -5,6 +5,7 @@ import com.mongodb.client.MongoDatabase;
 import spark.Request;
 import spark.Response;
 import umm3601.user.UserController;
+import umm3601.user.UserRequestHandler;
 
 import java.io.IOException;
 
@@ -22,6 +23,7 @@ public class Server {
         MongoDatabase userDatabase = mongoClient.getDatabase(userDatabaseName);
 
         UserController userController = new UserController(userDatabase);
+        UserRequestHandler userRequestHandler = new UserRequestHandler(userController);
 
         //Configure Spark
         port(serverPort);
@@ -61,9 +63,9 @@ public class Server {
 
         //List users, filtered using query parameters
 
-        get("api/users", userController::getUsers);
-        get("api/users/:id", userController::getUserJSON);
-        post("api/users/new", userController::addNewUser);
+        get("api/users", userRequestHandler::getUsers);
+        get("api/users/:id", userRequestHandler::getUserJSON);
+        post("api/users/new", userRequestHandler::addNewUser);
 
         // An example of throwing an unhandled exception so you can see how the
         // Java Spark debugger displays errors like this.
