+++ server/src/main/java/umm3601/user/UserRequestHandler.java
@@ -1,0 +1,108 @@
+package umm3601.user;
+
+import com.mongodb.BasicDBObject;
+import com.mongodb.util.JSON;
+import spark.Request;
+import spark.Response;
+
+/**
+ * Created by Brian on 11/29/2017.
+ */
+public class UserRequestHandler {
+
+    private final UserController userController;
+    public UserRequestHandler(UserController userController){
+        this.userController = userController;
+    }
+    /**
+     * Get a JSON response with a list of all the users in the database.
+     *
+     * @param req the HTTP request
+     * @param res the HTTP response
+     * @return one user in JSON formatted string and if it fails it will return text with a different HTTP status code
+     */
+    public String getUserJSON(Request req, Response res){
+        res.type("application/json");
+        String id = req.params("id");
+        String user;
+        try {
+            user = userController.getUserJSON(id);
+        } catch (IllegalArgumentException e) {
+            // This is thrown if the ID doesn't have the appropriate
+            // form for a Mongo Object ID.
+            // https://docs.mongodb.com/manual/reference/method/ObjectId/
+            res.status(400);
+            res.body("The requested user id " + id + " wasn't a legal Mongo Object ID.\n" +
+                "See 'https://docs.mongodb.com/manual/reference/method/ObjectId/' for more info.");
+            return "";
+        }
+        if (user != null) {
+            return user;
+        } else {
+            res.status(404);
+            res.body("The requested user with id " + id + " was not found");
+            return "";
+        }
+    }
+
+
+
+    /**
+     *@param req the HTTP request
+     * @param res the HTTP response
+     * @return an array of users in JSON formatted String
+     */
+    public String getUsers(Request req, Response res)
+    {
+        res.type("application/json");
+        return userController.getUsers(req.queryMap().toMap());
+    }
+
+
+    /**
+     *
+     * @param req the HTTP request
+     * @param res the HTTP response
+     * @return a boolean as whether the user was added succesfully or not
+     */
+    public boolean addNewUser(Request req, Response res)
+    {
+
+        res.type("application/json");
+        Object o = JSON.parse(req.body());
+        try {
+            if(o.getClass().equals(BasicDBObject.class))
+            {
+                try {
+                    BasicDBObject dbO = (BasicDBObject) o;
+
+                    String name = dbO.getString("name");
+                    //For some reason age is a string right now, caused by angular.
+                    //This is a problem and should not be this way but here ya go
+                    int age = dbO.getInt("age");
+                    String company = dbO.getString("company");
+                    String email = dbO.getString("email");
+
+                    System.err.println("Adding new user [name=" + name + ", age=" + age + " company=" + company + " email=" + email + ']');
+                    return userController.addNewUser(name, age, company, email);
+                }
+                catch(NullPointerException e)
+                {
+                    System.err.println("A value was malformed or omitted, new user request failed.");
+                    return false;
+                }
+
+            }
+            else
+            {
+                System.err.println("Expected BasicDBObject, received " + o.getClass());
+                return false;
+            }
+        }
+        catch(RuntimeException ree)
+        {
+            ree.printStackTrace();
+            return false;
+        }
+    }
+}
